requirements.txt
# LiveKit and Voice Processing
livekit==0.17.3
livekit-agents==0.9.6
livekit-plugins-openai==0.8.3
livekit-plugins-silero==0.6.3

# OpenAI
openai>=1.12.0

# Vector Database
faiss-cpu==1.7.4
numpy==1.24.3

# Data Processing
pandas==2.0.3
scikit-learn==1.3.0

# Environment and Configuration
python-dotenv==1.0.0

# Web Framework (for serving UI)
flask==3.0.0
flask-cors==4.0.0

# Utilities
aiohttp==3.9.1
pydantic==2.5.0
